add refreshtoken logic

+++ services/cros-jobs/emailTriggers.js
@@ -3,33 +3,49 @@ const redis = require("redis");
 const  { client }  = require('../../services/redisConnect');
 const { generateReport } = require("../email/reportService");
 
+let isRunning = false;
+
 async function processPendingReports() {
 
-  const patients = await client.sMembers("pending_reports");
+  if (isRunning) {
+    console.log("Previous report cycle still running");
+    return;
+  }
 
-  console.log("Pending patients:", patients);
+  isRunning = true;
 
-  for (const patientId of patients) {
+  try {
 
-    try {
+    const patients = await client.sMembers("pending_reports");
 
-      const key = `pending_reports:${patientId}`;
-      const fhirIds = await client.sMembers(key);
+    console.log("Pending patients:", patients);
 
-      console.log(`Processing ${patientId}`, fhirIds);
+    for (const patientId of patients) {
 
-      await generateReport(patientId, fhirIds);
+      try {
 
-      // cleanup after success
-      await client.del(key);
-      await client.sRem("pending_reports", patientId);
+        const key = `pending_reports:${patientId}`;
+        const fhirIds = await client.sMembers(key);
 
-    } catch (err) {
+        console.log(`Processing ${patientId}`, fhirIds);
 
-      console.error("Report generation failed", err);
+        await generateReport(patientId, fhirIds);
 
+        await client.del(key);
+        await client.sRem("pending_reports", patientId);
+
+      } catch (err) {
+
+        console.error(
+          `Report generation failed for ${patientId}`,
+          err.message
+        );
+
+      }
     }
 
+  } finally {
+    isRunning = false;
   }
 }
